requirements.txt
# 기본
pandas
numpy
requests
python-dotenv
aiohttp
schedule
tqdm

# 거래소 및 API 연동
ccxt
pybit==3.2.0
websocket-client

# AI/NLP 관련
openai
transformers
scikit-learn

# 한국어 형태소 분석
konlpy
JPype1

# 시각화 및 대시보드
streamlit
matplotlib
seaborn
plotly

# 기타 유틸
beautifulsoup4
feedparser
regex

# 백업 및 패키징
gitpython
pyinstaller

# 딥러닝 관련 (옵션)
tensorflow
torch
